% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anagrams.R
\name{get_anagrams}
\alias{get_anagrams}
\title{anagram function}
\usage{
get_anagrams(word, dictionary = "scrabble")
}
\arguments{
\item{word}{a word (or vector of length one with multiple characters/digits)}

\item{dictionary}{which dictionary should be used to look up words?  current options are 'scrabble' for the official scrabble dictionary and 'words' for a more expansive dictionary (including e.g. "mr").}
}
\description{
This function returns all possible words within a word, scrabble style
}
\examples{
perms <- get_anagrams("perms")
length(perms)
perms
}
\keyword{anagrams}
