% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{permutation function}
\usage{
permute(word)
}
\arguments{
\item{word}{a word (or vector of length one with multiple characters/digits)}
}
\description{
This function uses Heap's algorithm to obtain all permutations of a word
}
\examples{
perms <- permute("perms")
length(perms)
head(perms)
}
\keyword{permute}
