% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\docType{data}
\name{scrabble}
\alias{scrabble}
\title{All words that appear in the scrabble dictionary}
\format{A character vector with 178692 elements.}
\usage{
scrabble
}
\description{
A dataset containing all words in the scrabble dictionary in lowercase.
This is the recommended dataset for looking up words in `get_anagrams`
}
\keyword{datasets}
