% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{swap function}
\usage{
swap(x, from, to)
}
\arguments{
\item{x}{a vector}

\item{from}{the first element to swap}

\item{to}{the second element to swap}
}
\description{
This function allows you to swap the position of two elements of a vector
}
\examples{
x <- c(1, 2, 3)
swap(x, 1, 3)
}
\keyword{swap}
