% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{words}
\alias{words}
\title{Standard words file on Unix operating systems.}
\format{A character vector with 261448 elements.}
\usage{
words
}
\description{
A dataset containing all words in the starndard words unix file dictionary in lowercase.
This dataset is somewhat less strict than `scrabble`. (e.g., including 'mr' and 'mrs').
}
\keyword{datasets}
